requirements.txt
openai
feedparser
schedule
python-dotenv
beautifulsoup4
requests